pyproject.toml
[tool.poetry]
name = "bd-course"
version = "0.1.0"
description = ""
authors = ["Dmitrii <dimamedv1@gmail.com>"]
readme = "README.md"
packages = [{include = "bd_course"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
pytils = "^0.4.1"
pillow = "^10.1.0"
python-slugify = "^8.0.1"
django-cron = "^0.6.0"
yadisk = "^2.1.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^69.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
